test(quantize): move the wide differential where CI can gate it
The differential landed in the width-sharded integration suite, which
builds one tier at a time. A single-tier feature forwards only
decimal-scaled/<tier>, and the rescale matcher's baked-reciprocal arm
is behind x-wide / xx-wide, so every shard compiled that arm out: the
cells the tests were written for were covered but never gated.

The root crate's full-feature leg builds with wide,x-wide,xx-wide, so a
root test gated on x-wide sits exactly where all three of the matcher's
kernels are reachable. Same tests, one copy, now gated rather than
merely present. The D462 shift ladder walks the three arms in one run:
30 single-chunk, 100 chained below the threshold, 361 inside it.

The integration suite's quantize file keeps a pointer to where the wide
coverage went and why.
