+++ decimal-scale-test/tests/api/quantize.rs
@@ -3,6 +3,14 @@
 //!
 //! The deprecated `rescale` / `rescale_with` aliases (removed in 0.6.0)
 //! are covered by the main crate's `tests/deprecated_aliases.rs`.
+//!
+//! The wide tiers are covered by the main crate's
+//! `tests/wide_quantize_routing.rs`, not here. The scale-down's `/ 10^shift`
+//! routes through the rescale matcher, whose baked-reciprocal kernel exists
+//! only under `x-wide` / `xx-wide` — and this suite is built one tier at a
+//! time, which does not enable them. A root-crate test sits in the
+//! full-feature build where all three of that matcher's kernels are
+//! reachable.
 
 mod from_quantize {
     //! `D38::quantize` / `quantize_with` integration tests. Moved out of
@@ -394,207 +402,4 @@ mod from_rescale_modes {
         let v = D76::<0>::MAX;
         let _: D76<75> = v.quantize::<75>();
     }
-}
-
-mod wide_matcher_differential {
-    //! Wide-tier `quantize` scale-down, differenced against an independent
-    //! implementation of the same operation.
-    //!
-    //! # Why this module exists
-    //!
-    //! `quantize_with`'s scale-down routes its `/ 10^shift` through the
-    //! crate's rescale matcher, which picks between three kernels on
-    //! `(shift, significant width)`. Some of those cells were unreachable
-    //! from `quantize` at any width the rest of this file covers: the
-    //! widest tier tested elsewhere here is D76, whose storage is four u64
-    //! limbs, and the matcher's baked-reciprocal arm engages only from
-    //! twenty-four limbs upward with `shift` in `200..=1850`. Without these
-    //! tests a routing change at D462 and wider is invisible.
-    //!
-    //! # The oracle
-    //!
-    //! Not a hand-computed expected value — at these widths that would be a
-    //! second implementation with no reviewer. Each result is instead
-    //! differenced against `convert_from_with` at the *same* width, which
-    //! reaches the same operation by a wholly separate route: a typed
-    //! `div_rem` against `10^shift` in the integer layer, never touching
-    //! the rescale matcher. The two share only the mode decider, so they
-    //! disagree if the quotient, the remainder comparison, the last decimal
-    //! digit, or the sign handling differs on either side.
-    //!
-    //! # Both liveness guards matter
-    //!
-    //! A differential that only ever hits the zero-remainder early return
-    //! agrees trivially and proves nothing, so each test asserts that the
-    //! rounding actually ran:
-    //!
-    //! - the dense case asserts `Trunc` and `AwayFromZero` disagree, which
-    //!   holds exactly when the discarded digits are non-zero;
-    //! - the tie case asserts `HalfTowardZero` and `HalfAwayFromZero`
-    //!   disagree, which holds exactly when the remainder is the half —
-    //!   i.e. that the tie the test was built to create really is one.
-    //!
-    //! Both inputs are built near the storage maximum on purpose: the
-    //! matcher keys on the significant limb length after leading-zero
-    //! trimming, so a small value at a wide tier routes as a narrow one and
-    //! would silently miss the band. Each test asserts its input stayed
-    //! above `MAX >> 3` so that property cannot rot.
-    //!
-    //! # Which kernel these actually reach, and where they do not
-    //!
-    //! The gates below are per-tier, matching this suite's convention and
-    //! CI's width-sharded matrix. A single-tier feature (`d462` and its
-    //! siblings) forwards only `decimal-scaled/d462`, and the matcher's
-    //! baked-reciprocal arm sits behind `x-wide` / `xx-wide` — so in a
-    //! single-tier build every cell here resolves to one of the two
-    //! magic-number kernels, which is what CI exercises today. The same
-    //! tests route the `200..=1850` cells to the baked-reciprocal arm
-    //! automatically under any `x-wide` or `xx-wide` build, with no edit;
-    //! that arm is therefore covered by these tests but not, as the shards
-    //! currently stand, gated by CI.
-
-    use decimal_scaled::{Int, RoundingMode};
-
-    const ALL_MODES: [RoundingMode; 8] = [
-        RoundingMode::HalfToEven,
-        RoundingMode::HalfAwayFromZero,
-        RoundingMode::HalfTowardZero,
-        RoundingMode::Trunc,
-        RoundingMode::Floor,
-        RoundingMode::Ceiling,
-        RoundingMode::AwayFromZero,
-        RoundingMode::ZeroFiveUp,
-    ];
-
-    /// Emits the dense pair and the exact-tie pair for one
-    /// `(tier, limbs, source scale, target scale)` cell.
-    macro_rules! wide_quantize_differential {
-        ($dense:ident, $tie:ident, $ty:ident, $n:literal, $src:literal, $tgt:literal) => {
-            /// Generic (non-tie) remainders across all eight modes, both
-            /// signs, at three near-maximal magnitudes.
-            #[test]
-            fn $dense() {
-                type Src = decimal_scaled::$ty<$src>;
-                type Dst = decimal_scaled::$ty<$tgt>;
-
-                let one = Int::<$n>::ONE;
-                let three = one + one + one;
-                let seven = three + three + one;
-                let width_floor = Int::<$n>::MAX >> 3u32;
-
-                for mag in [Int::<$n>::MAX, Int::<$n>::MAX / three, Int::<$n>::MAX / seven] {
-                    assert!(
-                        mag >= width_floor,
-                        "{}: input dropped below the width floor, so it no longer \
-                         routes as a full-width value",
-                        stringify!($dense)
-                    );
-                    for raw in [mag, Int::<$n>::ZERO - mag] {
-                        let x = Src::from_bits(raw);
-
-                        let trunc = x.quantize_with::<$tgt>(RoundingMode::Trunc);
-                        let away = x.quantize_with::<$tgt>(RoundingMode::AwayFromZero);
-                        assert_ne!(
-                            trunc.to_bits(),
-                            away.to_bits(),
-                            "{}: discarded digits were zero, so the rounding path \
-                             never ran and this proves nothing",
-                            stringify!($dense)
-                        );
-
-                        for &mode in &ALL_MODES {
-                            let got = x.quantize_with::<$tgt>(mode);
-                            let want = Dst::convert_from_with(x, mode)
-                                .expect("a same-width scale-down cannot overflow");
-                            assert_eq!(
-                                got.to_bits(),
-                                want.to_bits(),
-                                "{}: quantize disagrees with convert_from_with under {:?}",
-                                stringify!($dense),
-                                mode
-                            );
-                        }
-                    }
-                }
-            }
-
-            /// The exact half-way remainder — the one input that separates
-            /// the three nearest modes — built by construction rather than
-            /// searched for.
-            #[test]
-            fn $tie() {
-                type Src = decimal_scaled::$ty<$src>;
-                type Dst = decimal_scaled::$ty<$tgt>;
-
-                let one = Int::<$n>::ONE;
-                let pow = Int::<$n>::TEN.pow(($src - $tgt) as u32);
-                // `10^k` is even for every `k >= 1`, so the half is exact.
-                let half = pow >> 1u32;
-                // The largest quotient leaving room for the half digit and
-                // for a bump away from zero.
-                let quotient = Int::<$n>::MAX / pow - one;
-                let raw = quotient * pow + half;
-
-                assert!(
-                    raw >= Int::<$n>::MAX >> 3u32,
-                    "{}: input dropped below the width floor, so it no longer \
-                     routes as a full-width value",
-                    stringify!($tie)
-                );
-
-                for signed in [raw, Int::<$n>::ZERO - raw] {
-                    let x = Src::from_bits(signed);
-
-                    let toward = x.quantize_with::<$tgt>(RoundingMode::HalfTowardZero);
-                    let away = x.quantize_with::<$tgt>(RoundingMode::HalfAwayFromZero);
-                    assert_ne!(
-                        toward.to_bits(),
-                        away.to_bits(),
-                        "{}: the remainder was not the exact half, so the tie-break \
-                         branch never ran",
-                        stringify!($tie)
-                    );
-
-                    for &mode in &ALL_MODES {
-                        let got = x.quantize_with::<$tgt>(mode);
-                        let want = Dst::convert_from_with(x, mode)
-                            .expect("a same-width scale-down cannot overflow");
-                        assert_eq!(
-                            got.to_bits(),
-                            want.to_bits(),
-                            "{}: quantize disagrees with convert_from_with under {:?}",
-                            stringify!($tie),
-                            mode
-                        );
-                    }
-                }
-            }
-        };
-    }
-
-    // D462 — twenty-four u64 limbs, the matcher's lower width bound. The
-    // three cells walk the shift axis across all three kernels: 30 is the
-    // single-chunk band, 100 the chained band below the baked-reciprocal
-    // threshold, 361 inside it.
-    #[cfg(feature = "d462")]
-    wide_quantize_differential!(d462_shift30_dense, d462_shift30_tie, D462, 24, 130, 100);
-    #[cfg(feature = "d462")]
-    wide_quantize_differential!(d462_shift100_dense, d462_shift100_tie, D462, 24, 200, 100);
-    #[cfg(feature = "d462")]
-    wide_quantize_differential!(d462_shift361_dense, d462_shift361_tie, D462, 24, 461, 100);
-
-    // The wider tiers inside the baked-reciprocal band.
-    #[cfg(feature = "d616")]
-    wide_quantize_differential!(d616_shift500_dense, d616_shift500_tie, D616, 32, 600, 100);
-    #[cfg(feature = "d924")]
-    wide_quantize_differential!(d924_shift800_dense, d924_shift800_tie, D924, 48, 900, 100);
-    #[cfg(feature = "d1232")]
-    wide_quantize_differential!(
-        d1232_shift1100_dense,
-        d1232_shift1100_tie,
-        D1232,
-        64,
-        1200,
-        100
-    );
 }
